SERVERPROPERTY, SESSIONPROPERTY, CONNECTIONPROPERTY, COLLATIONPROPERTY, LOGINPROPERTY, SESSION_CONTEXT, SQL_VARIANT_PROPERTY and sys.configurations' value columns project true sql_variant instead of inner-type substitution, carrying per-property inner base types (COLLATIONPROPERTY Version is tinyint, sys.configurations values are int) so mixed-inner UNION/ISNULL/COALESCE/CASE stay variant, SELECT INTO creates a sql_variant column, SESSION_CONTEXT preserves the stored value's type, comparisons unwrap, the wire reports 0x62 COLMETADATA, and arithmetic rejects with Msg 257 (variant with a convertible non-variant) split enforced identically at parse and run time.

+++ SqlServerSimulator/BuiltInResources.LegacyCompat.cs
@@ -144,7 +144,7 @@ internal static partial class BuiltInResources
             var status = (short)((isDynamic ? 1 : 0) + (isAdvanced ? 2 : 0));
             rows[i] =
             [
-                SqlValue.FromInt32((int)value),
+                SqlValue.FromInt32(value),
                 SqlValue.FromInt32(id),
                 SqlValue.FromNVarchar(description),
                 SqlValue.FromInt16(status),
